Fix envvar-arg mapping when spaces are in value.
The environment variable to ES option mapping in `es-docker` does not
work correctly when an option value contains whitespace.  This commit
changes the assembly of options to use an array instead of a string, and
then uses a proper all-array-items-to-words expansion at the end so
things like quoting don't need to be worried about.